SyncFit/requirements.txt
pandas
faker
scikit-learn
xgboost
streamlit
joblib
apache-airflow
